Move count / study fetching down into View Data, Request Study
so that we can get this out of Vuex.
