+++ lib/api/WebApi.js
@@ -3,6 +3,8 @@ import {
 } from '@/lib/Constants';
 import BackendClient from '@/lib/api/BackendClient';
 
+const MAX_PER_CATEGORY = 10;
+
 const apiClient = new BackendClient('/api');
 
 async function deleteStudyRequests(csrf, isSupervisor, studyRequests) {
@@ -17,6 +19,32 @@ async function deleteStudyRequests(csrf, isSupervisor, studyRequests) {
     ({ id }) => apiClient.fetch(`/requests/study/${id}`, options),
   );
   await Promise.all(promisesStudyRequests);
+}
+async function getCountsByCentreline(feature) {
+  const { centrelineId, centrelineType } = feature;
+
+  const dataCounts = {
+    centrelineId,
+    centrelineType,
+    maxPerCategory: MAX_PER_CATEGORY,
+  };
+  const optionsCounts = { data: dataCounts };
+
+  const dataStudies = {
+    centrelineId,
+    centrelineType,
+  };
+  const optionsStudies = { data: dataStudies };
+
+  const [
+    { counts },
+    studies,
+  ] = await Promise.all([
+    apiClient.fetch('/counts/byCentreline', optionsCounts),
+    apiClient.fetch('/studies/byCentreline', optionsStudies),
+  ]);
+
+  return { counts, studies };
 }
 
 async function getLocationsByFeature(centrelineTypesAndIds) {
@@ -111,6 +139,7 @@ async function putStudyRequests(csrf, isSupervisor, studyRequests) {
 
 const WebApi = {
   deleteStudyRequests,
+  getCountsByCentreline,
   getLocationByFeature,
   getLocationsByFeature,
   getUserStudyRequests,
@@ -121,6 +150,7 @@ const WebApi = {
 export {
   WebApi as default,
   deleteStudyRequests,
+  getCountsByCentreline,
   getLocationByFeature,
   getLocationsByFeature,
   getUserStudyRequests,
