+++ web/views/FcRequestStudyRequest.vue
@@ -30,6 +30,9 @@
 <script>
 import { mapGetters, mapMutations, mapState } from 'vuex';
 
+import {
+  getCountsByCentreline,
+} from '@/lib/api/WebApi';
 import FcCardTableStudiesRequested from '@/web/components/FcCardTableStudiesRequested.vue';
 import TdsActionDropdown from '@/web/components/tds/TdsActionDropdown.vue';
 import ArrayUtils from '@/lib/ArrayUtils';
@@ -79,6 +82,12 @@ export default {
     FcCardTableStudiesRequested,
     TdsActionDropdown,
   },
+  data() {
+    return {
+      counts: [],
+      studies: [],
+    };
+  },
   computed: {
     items() {
       return this.studyRequest.studies.map(({ studyType }, id) => {
@@ -90,10 +99,12 @@ export default {
       'studyTypesWithWarnings',
     ]),
     ...mapState('requestStudy', ['studyRequest']),
-    ...mapState([
-      'counts',
-      'studies',
-    ]),
+    ...mapState(['location']),
+  },
+  beforeRouteEnter(to, from, next) {
+    next((vm) => {
+      vm.syncFromRoute(to);
+    });
   },
   methods: {
     onAddStudy(studyType) {
@@ -110,6 +121,15 @@ export default {
     },
     onRemoveStudy(item) {
       this.removeStudyFromStudyRequest(item.id);
+    },
+    async syncFromRoute() {
+      const { centrelineId, centrelineType } = this.location;
+      const { counts, studies } = await getCountsByCentreline({
+        centrelineId,
+        centrelineType,
+      });
+      this.counts = counts;
+      this.studies = studies;
     },
     ...mapMutations('requestStudy', [
       'addStudyToStudyRequest',
